添加弹框，删除时弹出

+++ public/js/main.js
@@ -24,15 +24,27 @@ const vm = new Vue({
         this.newTitle = '';
       },
       removeData: function (id) {
-        let _this = this;
-        this.$http.delete(this.apiRequestUrl + '/' + id).then(function () {
-          _this.$http.get(_this.apiRequestUrl).then(function (res) {
-            this.lists = res.data;
-            let len = res.data.length;
-            let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
-            this.id = dataId ? parseInt(dataId) + 1 : 0;
-          })
-        });
+        swal({
+          title            : '确定要删除吗？',
+          type             : 'warning',
+          showCancelButton : true,
+          confirmButtonText: '确定',
+          cancelButtonText : '取消'
+        }).then((result) => {
+          if (result.value) {
+            let _this = this;
+            this.$http.delete(this.apiRequestUrl + '/' + id).then(function () {
+              _this.$http.get(_this.apiRequestUrl).then(function (res) {
+                this.lists = res.data;
+                let len = res.data.length;
+                let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
+                this.id = dataId ? parseInt(dataId) + 1 : 0;
+              })
+            }).then(function () {
+              swal('删除成功！')
+            });
+          }
+        })
       }
       ,
       editData  : function (id) {
@@ -74,6 +86,9 @@ const vm = new Vue({
             _this.id = dataId ? parseInt(dataId) + 1 : 0;
           })
         });
+      },
+      cancelEdit:function () {
+        this.isEdit = false;
       }
 
     }
